Added custom calendar creation

+++ calendar.rb
@@ -1,7 +1,9 @@
 require_relative "events.rb"
+require_relative "calendar_creator.rb"
+require 'byebug'
 
 def calendar
-    if File.file?("date.txt") == true
+    if File.file?("date.txt")
         prev_date = File.readlines("date.txt")
         day = prev_date[0].chomp.to_i
         month = prev_date[1].chomp.to_i
@@ -18,17 +20,40 @@ def calendar
         days_passed = STDIN.gets.chomp.to_i
     end
 
+    if File.file?("calendar.txt") != true
+        calendar_creator()
+    end
 
-    days = ["1st", "2nd", "3rd", "4th", "5th", "6th", "7th", "8th", "9th", "10th", "11th", "12th", "13th", "14th", "15th", "16th", "17th", "18th", "19th", "20th", "21th", "22th", "23th", "24th", "25th", "26th", "27th", "28th", "29th", "30th"]
-    months = ["January", "Febuary", "Mars", "April", "May", "June", "July", "August", "September", "October", "November", "December"]
+    calendar_arr = File.readlines("calendar.txt")
+    daynum = calendar_arr[0].to_i
+    days = []
+    i = 1
+    while i <= daynum
+        if "#{i}"[-1] != "1" && "#{i}"[-1] != "2" && "#{i}"[-1] != "3"
+            days << "#{i}th"
+        elsif "#{i}"[-2] == "1"
+            days << "#{i}th"
+        elsif "#{i}"[-1] == "1"
+            days << "#{i}st"
+        elsif "#{i}"[-1] == "2"
+            days << "#{i}nd"
+        elsif "#{i}"[-1] == "3"
+            days << "#{i}rd"
+        else
+            return "WTF are you doing?"
+        end
+        i += 1
+    end
+    months = calendar_arr[1].split(", ")
+
     output = "#{days[day]} of #{months[month]}, #{year}"
     output += events(day, month, year)
-    output += lunar_phases(day, month, year)
+    output += lunar_phases(day, month, year, daynum, months.length)
     output += "\n#{days_passed} days have passed since the start of the campaign."
     puts output
 
     continue = STDIN.gets.chomp
-    while continue != "save" && continue != "Save" && continue != "MAXIMUM RESET"
+    while continue != "save" && continue != "Save" && continue != "MAXIMUM RESET" && continue != "calendar reset"
         if day < days.length - 1
             day += 1
         elsif month < months.length - 1
@@ -42,14 +67,25 @@ def calendar
         days_passed += 1
         output = "#{days[day]} of #{months[month]}, #{year}"
         output += events(day, month, year)
-        output += lunar_phases(day, month, year)
+        output += lunar_phases(day, month, year, daynum, months.length)
         output += "\n#{days_passed} days have passed since the start of the campaign."
         puts output
         continue = STDIN.gets.chomp
     end
     if continue == "MAXIMUM RESET"
-        File.delete("date.txt")
-        return "It is done..."
+        if File.file?("date.txt")
+            File.delete("date.txt")
+            return "It is done..."
+        else
+            return "You fool! There is nothing to reset!"
+        end
+    elsif continue == "calendar reset"
+        if File.file?("calendar.txt")
+            File.delete("calendar.txt")
+            return "It is done..."
+        else
+            return "You fool! There is no calendar to reset!"
+        end
     end
     data = "#{day}\n#{month}\n#{year}\n#{days_passed}"
     File.write("date.txt", data)
